Split Vulkan backend source code into several files
