+++ Src/Render/RHI/Backends/Vulkan/VulkanDevice.h
@@ -7,29 +7,8 @@
 #include <limits>
 #include <vector>
 
-#if defined(_WIN32)
-#ifndef VK_USE_PLATFORM_WIN32_KHR
-#define VK_USE_PLATFORM_WIN32_KHR
-#endif
-#elif defined(__linux__)
-#if defined(GLAB_GLFW_X11_NATIVE)
-#ifndef VK_USE_PLATFORM_XLIB_KHR
-#define VK_USE_PLATFORM_XLIB_KHR
-#endif
-#endif
-#if defined(GLAB_GLFW_WAYLAND_NATIVE)
-#ifndef VK_USE_PLATFORM_WAYLAND_KHR
-#define VK_USE_PLATFORM_WAYLAND_KHR
-#endif
-#endif
-#endif
-
-#include <volk/volk.h>
-
 #include "Render/RHI/Backends/Common/RHIShellCommon.h"
-
-struct VmaAllocator_T;
-struct VmaAllocation_T;
+#include "Render/RHI/Backends/Vulkan/VulkanCommon.h"
 
 class VulkanDevice;
 class VulkanSwapchainTexture;
